requirements.txt
Django>=5.1.3
djangorestframework>=3.15.2
django-filter>=24.3
django-cors-headers